Fix the TxModal Finalization status
Fix the TxModal Finalization status. The current code shows 'N/A' when the status is 'pending' and when the status is 'success', the transaction object still has a pending object so the current check is always in a pending status. So I changed the code to just check the status of the transaction.